backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile .\backend\requirements.in -o .\backend\requirements.txt
aiohttp==3.9.3
    # via aiohttp-middlewares
aiohttp-middlewares==2.3.0
aiosignal==1.3.1
    # via aiohttp
aiosqlite==0.20.0
annotated-types==0.6.0
    # via pydantic
async-timeout==4.0.3
    # via aiohttp-middlewares
asyncpg==0.29.0
attrs==23.2.0
    # via aiohttp
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.0.3
    # via sqlalchemy
idna==3.6
    # via yarl
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
pillow==10.2.0
pydantic==2.6.1
pydantic-core==2.16.2
    # via pydantic
sqlalchemy==2.0.27
typing-extensions==4.9.0
    # via
    #   aiosqlite
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
yarl==1.9.4
    # via
    #   aiohttp
    #   aiohttp-middlewares
